board,삭제버튼로직 수정

+++ src/pages/Board.jsx
@@ -1,22 +1,55 @@
 import { useEffect, useState } from "react";
-
+import { Link } from "react-router-dom";
 function Board() {
   const [boards, setBoards] = useState([]);
+  const [loginUser,setLoginUser] = useState("");
+
+  const fetchBoards = async()=>{
+    const res = await fetch("https://ani-5.onrender.com/api/board", {
+      credentials: "include",
+    });
+    const data = await res.json();
+    setBoards(data);
+  };
 
   useEffect(() => {
-    fetch("https://ani-5.onrender.com/api/board", {
-      credentials: "include",
+  fetchBoards();
+
+  fetch("https://ani-5.onrender.com/api/user/me", {
+    credentials: "include",
+  })
+    .then((res) => {
+      if (!res.ok) {
+        return null;
+      }
+      return res.json();
     })
-      .then((res) => res.json())
-      .then((data) => setBoards(data));
-  }, []);
+    .then((data) => {
+      if (data) {
+        setLoginUser(data.username);
+      }
+    })
+    .catch((err) => {
+      console.error("유저정보부르기 실패", err);
+    });
+}, []);
 
   const handleDelete = async (id) => {
-    await fetch(`https://ani-5.onrender.com/api/board/${id}`, {
+    if(!window.confirm("진짜 삭제하노?")) return;
+
+    const res = await fetch(`https://ani-5.onrender.com/api/board/${id}`, {
       method: "DELETE",
       credentials: "include",
     });
-    setBoards(boards.filter((b) => b.id !== id));
+
+    if (res.status === 403) {
+      const msg = await res.text();
+      alert(msg);
+      return;
+    }
+
+    alert("삭제 완료");
+    fetchBoards();
   };
 
   return (
@@ -53,11 +86,15 @@ function Board() {
             <tr key={board.id}>
               <td>{board.id}</td>
               <td>
-                <a href={`/board/${board.id}`}>{board.title}</a>
+                <Link to={`/board/${board.id}`}>{board.title}</Link>
               </td>
               <td>{board.user.username}</td>
               <td>
-                <button onClick={() => handleDelete(board.id)}>삭제</button>
+                {board.user.username===loginUser &&(
+                  <button onClick={() => handleDelete(board.id)}>
+                    삭제
+                  </button>
+                )}
               </td>
             </tr>
           ))}
